Added elastic feature

+++ utils/tower_adaptation.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-ln -s /vagrant /var/lib/awx/projects/ansible-demo
+[[ -L /var/lib/awx/projects/ansible-demo ]] || ln -s /vagrant /var/lib/awx/projects/ansible-demo
 chown -h awx.awx /var/lib/awx/projects/ansible-demo
 sed -i s/#host_key_checking/host_key_checking/ /etc/ansible/ansible.cfg
 iptables -F
